+++ src/connectedvmware/azext_connectedvmware/vendored_sdks/connectedvmware/operations/_machine_extensions_operations.py
@@ -242,7 +242,7 @@ class MachineExtensionsOperations(object):
     instantiates it for you and attaches it as an attribute.
 
     :ivar models: Alias to model classes used in this operation group.
-    :type models: ~azure_arc_vmware_management_service_api.models
+    :type models: ~azure.mgmt.connectedvmware.models
     :param client: Client for service requests.
     :param config: Configuration of service client.
     :param serializer: An object model serializer.
@@ -335,7 +335,7 @@ class MachineExtensionsOperations(object):
         :param extension_name: The name of the machine extension.
         :type extension_name: str
         :param extension_parameters: Parameters supplied to the Create Machine Extension operation.
-        :type extension_parameters: ~azure_arc_vmware_management_service_api.models.MachineExtension
+        :type extension_parameters: ~azure.mgmt.connectedvmware.models.MachineExtension
         :keyword callable cls: A custom type or function that will be passed the direct response
         :keyword str continuation_token: A continuation token to restart a poller from a saved state.
         :keyword polling: By default, your polling method will be ARMPolling. Pass in False for this
@@ -346,8 +346,7 @@ class MachineExtensionsOperations(object):
          Retry-After header is present.
         :return: An instance of LROPoller that returns either MachineExtension or the result of
          cls(response)
-        :rtype:
-         ~azure.core.polling.LROPoller[~azure_arc_vmware_management_service_api.models.MachineExtension]
+        :rtype: ~azure.core.polling.LROPoller[~azure.mgmt.connectedvmware.models.MachineExtension]
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         api_version = kwargs.pop('api_version', "2023-03-01-preview")  # type: str
@@ -472,8 +471,7 @@ class MachineExtensionsOperations(object):
         :param extension_name: The name of the machine extension.
         :type extension_name: str
         :param extension_parameters: Parameters supplied to the Create Machine Extension operation.
-        :type extension_parameters:
-         ~azure_arc_vmware_management_service_api.models.MachineExtensionUpdate
+        :type extension_parameters: ~azure.mgmt.connectedvmware.models.MachineExtensionUpdate
         :keyword callable cls: A custom type or function that will be passed the direct response
         :keyword str continuation_token: A continuation token to restart a poller from a saved state.
         :keyword polling: By default, your polling method will be ARMPolling. Pass in False for this
@@ -484,8 +482,7 @@ class MachineExtensionsOperations(object):
          Retry-After header is present.
         :return: An instance of LROPoller that returns either MachineExtension or the result of
          cls(response)
-        :rtype:
-         ~azure.core.polling.LROPoller[~azure_arc_vmware_management_service_api.models.MachineExtension]
+        :rtype: ~azure.core.polling.LROPoller[~azure.mgmt.connectedvmware.models.MachineExtension]
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         api_version = kwargs.pop('api_version', "2023-03-01-preview")  # type: str
@@ -571,8 +568,13 @@ class MachineExtensionsOperations(object):
             map_error(status_code=response.status_code, response=response, error_map=error_map)
             raise HttpResponseError(response=response, error_format=ARMErrorFormat)
 
+        response_headers = {}
+        if response.status_code == 202:
+            response_headers['Location']=self._deserialize('str', response.headers.get('Location'))
+            
+
         if cls:
-            return cls(pipeline_response, None, {})
+            return cls(pipeline_response, None, response_headers)
 
     _delete_initial.metadata = {'url': "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ConnectedVMwarevSphere/virtualMachines/{virtualMachineName}/extensions/{extensionName}"}  # type: ignore
 
@@ -663,7 +665,7 @@ class MachineExtensionsOperations(object):
         :type extension_name: str
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: MachineExtension, or the result of cls(response)
-        :rtype: ~azure_arc_vmware_management_service_api.models.MachineExtension
+        :rtype: ~azure.mgmt.connectedvmware.models.MachineExtension
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         cls = kwargs.pop('cls', None)  # type: ClsType["_models.MachineExtension"]
@@ -729,7 +731,7 @@ class MachineExtensionsOperations(object):
         :return: An iterator like instance of either MachineExtensionsListResult or the result of
          cls(response)
         :rtype:
-         ~azure.core.paging.ItemPaged[~azure_arc_vmware_management_service_api.models.MachineExtensionsListResult]
+         ~azure.core.paging.ItemPaged[~azure.mgmt.connectedvmware.models.MachineExtensionsListResult]
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         api_version = kwargs.pop('api_version', "2023-03-01-preview")  # type: str