+++ src/spring/azext_spring/tests/latest/test_asa_scenario.py
@@ -161,3 +161,20 @@ class SslTests(ScenarioTest):
         self.assertTrue(len(app_result) > 0)
 
         self.cmd('spring app delete --name {app}  -g {rg} -s {serviceName}')
+
+
+class FlushVirtualNetworkDnsSettingTest(ScenarioTest):
+
+    @SpringResourceGroupPreparer(dev_setting_name=SpringTestEnvironmentEnum.STANDARD['resource_group_name'])
+    @SpringPreparer(**SpringTestEnvironmentEnum.STANDARD['spring'])
+    def test_flush_virtualnetwork_dns_settings_service(self, resource_group, spring):
+        self.kwargs.update({
+            'serviceName': spring,
+            'resource_group': resource_group,
+        })
+
+        self.cmd('spring flush-virtualnetwork-dns-settings -n {serviceName} -g {resource_group}')
+        self.cmd('spring show --name {serviceName} -g {resource_group}', checks=[
+            self.check('properties.provisioningState', 'Succeeded'),
+            self.check('properties.powerState', 'Running')
+        ])