+++ src/spring/azext_spring/vendored_sdks/appplatform/v2023_01_01_preview/models/_models_py3.py
@@ -3845,8 +3845,7 @@ class CustomContainer(_serialization.Model):
     :ivar image_registry_credential: Credential of the image registry.
     :vartype image_registry_credential:
      ~azure.mgmt.appplatform.v2023_01_01_preview.models.ImageRegistryCredential
-    :ivar language_framework: Language framework of the container image uploaded. Supported values:
-     "springboot", "", null.
+    :ivar language_framework: Language framework of the container image uploaded.
     :vartype language_framework: str
     """
 
@@ -3885,8 +3884,7 @@ class CustomContainer(_serialization.Model):
         :keyword image_registry_credential: Credential of the image registry.
         :paramtype image_registry_credential:
          ~azure.mgmt.appplatform.v2023_01_01_preview.models.ImageRegistryCredential
-        :keyword language_framework: Language framework of the container image uploaded. Supported
-         values: "springboot", "", null.
+        :keyword language_framework: Language framework of the container image uploaded.
         :paramtype language_framework: str
         """
         super().__init__(**kwargs)